clear the EtherShield of 2 functions, better be in the file ip_arp_udp_tcp and change the name of paquetweb (FR) to ES_ProcessWebPacket (EN)
