+++ inc/EtherShield.h
@@ -21,18 +21,11 @@
 #define ETHERSHIELD_H
 
 #include "defines.h"
-
-#define bool _Bool
-#define TRUE 1
-#define FALSE 0
-
-#include <inttypes.h>
 #include "enc28j60.h"
-#include "ip_arp_udp_tcp.h"
 #include "net.h"
 
 void ES_FullConnection();
-void paquetweb();
+void ES_ProcessWebPacket();
 void ES_enc28j60SpiInit( SPI_HandleTypeDef *hspi );
 void ES_enc28j60Init( uint8_t* macaddr);
 void ES_enc28j60clkout(uint8_t clk);
@@ -105,7 +98,6 @@ uint8_t *ES_dnslkup_getip( void );
 void ES_dnslkup_set_dnsip(uint8_t *dnsipaddr);
 void ES_dnslkup_request(uint8_t *buf, uint8_t *hoststr );
 uint8_t ES_udp_client_check_for_dns_answer(uint8_t *buf,uint16_t plen);
-uint8_t resolveHostname(uint8_t *buf, uint16_t buffer_size, uint8_t *hostname );
 #endif
 
 #ifdef DHCP_client
@@ -115,7 +107,6 @@ void ES_dhcp_start(uint8_t *buf, uint8_t *macaddrin, uint8_t *ipaddrin,
 		uint8_t *dnssvrin );
 
 uint8_t ES_check_for_dhcp_answer(uint8_t *buf,uint16_t plen);
-uint8_t allocateIPAddress(uint8_t *buf, uint16_t buffer_size, uint8_t *mymac, uint16_t myport, uint8_t *myip, uint8_t *mynetmask, uint8_t *gwip, uint8_t *dnsip, uint8_t *dhcpsvrip );
 #endif
 
 #define HTTP_HEADER_START ((uint16_t)TCP_SRC_PORT_H_P+(buf[TCP_HEADER_LEN_P]>>4)*4)
