+++ src/EtherShield.c
@@ -1,47 +1,60 @@
-#include "defines.h"
-#include "main.h"
-#include "enc28j60.h"
-#include "ip_arp_udp_tcp.h"
-#include "websrv_help_functions.h"
-
-#ifdef DNS_client
-	#include "dnslkup.h"
-#endif
-
-#ifdef DHCP_client
-	#include "dhcp.h"
-#endif
+/*
+ * This file contains the implementation for establishing a full connection using the EtherShield library
+ * with an ENC28J60 Ethernet module. It includes functions to initialize the connection, allocate IP addresses
+ * via DHCP, resolve DNS hostnames, and request NTP time updates. Additionally, it processes incoming packets
+ * for ICMP and TCP protocols.
+ */
 
 #include "EtherShield.h"
 
 #define BUFFER_SIZE 400
 
-uint8_t hostname[]= "www.google.com";
+uint8_t hostname[] = "www.google.com";
 
+// Function to initialize and establish a full connection
 void ES_FullConnection(SPI_HandleTypeDef *hspi) {
-	enc28j60_set_spi(hspi);
-	enc28j60Init(macaddrin);
-	enc28j60Write(ECOCON, 3 & 0x7);
-    enc28j60PhyWrite(PHLCON,0x3880);
-	enc28j60PhyWrite(PHLCON,0x3990);
-	enc28j60PhyWrite(PHLCON,0x3476);
-	uint8_t bufDHCP[BUFFER_SIZE];
-	allocateIPAddress(bufDHCP, BUFFER_SIZE, macaddrin, 80, ipaddrin, maskin, gwipin, dnssvrin, dhcpsvrin);
-	client_set_gwip(gwipin);
-	uint8_t bufDNS[BUFFER_SIZE];
-	dnslkup_request(bufDNS, domainName);
-	udp_client_check_for_dns_answer(bufDNS, plen);
-	dnslkup_set_dnsip(dnsipaddr);
-	resolveHostname(bufDNS, BUFFER_SIZE, hostname);
-	uint8_t bufNTP[BUFFER_SIZE];
-	client_ntp_request(bufNTP, ntpip, 123);
-	udpLog2("ES_FullConnection", "WORKS!");
+    // Set the SPI handle for the ENC28J60
+    enc28j60_set_spi(hspi);
+
+    // Initialize the ENC28J60 with the MAC address
+    enc28j60Init(macaddrin);
+
+    // Configure the ENC28J60 clock
+    enc28j60Write(ECOCON, 3 & 0x7);
+
+    // Configure the ENC28J60 PHY LEDs
+    enc28j60PhyWrite(PHLCON, 0x3880);
+    enc28j60PhyWrite(PHLCON, 0x3990);
+    enc28j60PhyWrite(PHLCON, 0x3476);
+
+    // Allocate IP address via DHCP
+    uint8_t bufDHCP[BUFFER_SIZE];
+    allocateIPAddress(bufDHCP, BUFFER_SIZE, macaddrin, 80, ipaddrin, maskin, gwipin, dnssvrin, dhcpsvrin);
+
+    // Set the gateway IP for the client
+    client_set_gwip(gwipin);
+
+    // Resolve the DNS hostname
+    uint8_t bufDNS[BUFFER_SIZE];
+    dnslkup_request(bufDNS, domainName);
+    udp_client_check_for_dns_answer(bufDNS, plen);
+    dnslkup_set_dnsip(dnsipaddr);
+    resolveHostname(bufDNS, BUFFER_SIZE, hostname);
+
+    // Send NTP request to get the current time
+    uint8_t bufNTP[BUFFER_SIZE];
+    client_ntp_request(bufNTP, ntpip, 123);
+
+    // Log the success of the full connection setup
+    udpLog2("ES_FullConnection", "WORKS!");
 }
 
-void paquetweb() {
-	static uint8_t packet[500];
-	packetloop_icmp_tcp(packet, enc28j60PacketReceive(500, packet));
+// Function to process incoming web packets
+void ES_ProcessWebPacket() {
+    static uint8_t packet[500];
+    packetloop_icmp_tcp(packet, enc28j60PacketReceive(500, packet));
 }
+
 /**
  * Initialise SPI, separate from main initialisation so that
  * multiple SPI devices can be used together
@@ -355,65 +368,6 @@ void ES_dnslkup_request(uint8_t *buf,uint8_t *hostname) {
 uint8_t ES_udp_client_check_for_dns_answer(uint8_t *buf,uint16_t plen){
 	return( udp_client_check_for_dns_answer( buf, plen) );
 }
-
-
-// Perform all processing to resolve a hostname to IP address.
-// Returns 1 for successful Name resolution, 0 otherwise
-uint8_t resolveHostname(uint8_t *buf, uint16_t buffer_size, uint8_t *hostname ) {
-  uint16_t dat_p;
-  int plen = 0;
-  long lastDnsRequest = HAL_GetTick();
-  uint8_t dns_state = DNS_STATE_INIT;
-  bool gotAddress = FALSE;
-  uint8_t dnsTries = 3;	// After 10 attempts fail gracefully so other action can be carried out
-
-  while( !gotAddress ) {
-    // handle ping and wait for a tcp packet
-    plen = enc28j60PacketReceive(buffer_size, buf);
-    dat_p=packetloop_icmp_tcp(buf,plen);
-
-    // We have a packet
-    // Check if IP data
-    if (dat_p == 0) {
-      if (client_waiting_gw() ) {
-        // No ARP received for gateway
-        continue;
-      }
-      // It has IP data
-      if (dns_state==DNS_STATE_INIT) {
-        dns_state=DNS_STATE_REQUESTED;
-        lastDnsRequest = HAL_GetTick();
-        dnslkup_request(buf,hostname);
-        continue;
-      }
-      if (dns_state!=DNS_STATE_ANSWER){
-        // retry every minute if dns-lookup failed:
-        if (HAL_GetTick() > (lastDnsRequest + 60000L) ){
-	  if( --dnsTries <= 0 ) 
-	    return 0;		// Failed to allocate address
-
-          dns_state=DNS_STATE_INIT;
-          lastDnsRequest = HAL_GetTick();
-        }
-        // don't try to use client before
-        // we have a result of dns-lookup
-
-        continue;
-      }
-    } 
-    else {
-      if (dns_state==DNS_STATE_REQUESTED && udp_client_check_for_dns_answer( buf, plen ) ){
-        dns_state=DNS_STATE_ANSWER;
-        //client_set_wwwip(dnslkup_getip());
-        client_tcp_set_serverip(dnslkup_getip());
-	gotAddress = TRUE;
-      }
-    }
-  }
-  
-  return 1;
-}
-
 #endif		// DNS_client
 
 #ifdef DHCP_client
@@ -429,61 +383,6 @@ uint8_t ES_dhcp_state(void)
 uint8_t ES_check_for_dhcp_answer(uint8_t *buf,uint16_t plen){
 	return( check_for_dhcp_answer( buf, plen) );
 }
-
-// Utility functions 
-
-// Perform all processing to get an IP address plus other addresses returned, e.g. gw, dns, dhcp server.
-// Returns 1 for successful IP address allocation, 0 otherwise
-uint8_t allocateIPAddress(uint8_t *buf, uint16_t buffer_size, uint8_t *mymac, uint16_t myport, uint8_t *myip, uint8_t *mynetmask, uint8_t *gwip, uint8_t *dnsip, uint8_t *dhcpsvrip ) {
-  uint16_t dat_p;
-  int plen = 0;
-  long lastDhcpRequest = HAL_GetTick();
-  uint8_t dhcpState = 0;
-  bool gotIp = FALSE;
-  uint8_t dhcpTries = 10;	// After 10 attempts fail gracefully so other action can be carried out
-
-  dhcp_start( buf, mymac, myip, mynetmask,gwip, dnsip, dhcpsvrip );
-
-  while( !gotIp ) {
-    // handle ping and wait for a tcp packet
-    plen = enc28j60PacketReceive(buffer_size, buf);
-    dat_p=packetloop_icmp_tcp(buf,plen);
-    if(dat_p==0) {
-      check_for_dhcp_answer( buf, plen);
-      dhcpState = dhcp_state();
-      // we are idle here
-      if( dhcpState != DHCP_STATE_OK ) {
-        if (HAL_GetTick() > (lastDhcpRequest + 10000L) ){
-          lastDhcpRequest = HAL_GetTick();
-	  if( --dhcpTries <= 0 ) 
-		  return 0;		// Failed to allocate address
-          // send dhcp
-          dhcp_start( buf, mymac, myip, mynetmask,gwip, dnsip, dhcpsvrip );
-        }
-      } else {
-        if( !gotIp ) {
-          gotIp = TRUE;
-
-          //init the ethernet/ip layer:
-          init_ip_arp_udp_tcp(mymac, myip, myport);
-
-          // Set the Router IP
-          client_set_gwip(gwip);  // e.g internal IP of dsl router
-
-#ifdef DNS_client
-          // Set the DNS server IP address if required, or use default
-          dnslkup_set_dnsip( dnsip );
-#endif
-
-        }
-      }
-    }
-  }
-
-  return 1;
-
-}
-
 #endif		// DHCP_client
 
 void ES_enc28j60PowerUp(){
