+++ src/ip_arp_udp_tcp.c
@@ -1386,6 +1386,117 @@ uint8_t packetloop_icmp_checkreply(uint8_t *buf,uint8_t *ip_monitoredhost)
 }
 #endif // PING_client
 
+// Utility functions
+
+// Perform all processing to get an IP address plus other addresses returned, e.g. gw, dns, dhcp server.
+// Returns 1 for successful IP address allocation, 0 otherwise
+uint8_t allocateIPAddress(uint8_t *buf, uint16_t buffer_size, uint8_t *mymac, uint16_t myport, uint8_t *myip, uint8_t *mynetmask, uint8_t *gwip, uint8_t *dnsip, uint8_t *dhcpsvrip ) {
+  uint16_t dat_p;
+  int plen = 0;
+  long lastDhcpRequest = HAL_GetTick();
+  uint8_t dhcpState = 0;
+  bool gotIp = FALSE;
+  uint8_t dhcpTries = 10;	// After 10 attempts fail gracefully so other action can be carried out
+
+  dhcp_start( buf, mymac, myip, mynetmask,gwip, dnsip, dhcpsvrip );
+
+  while( !gotIp ) {
+    // handle ping and wait for a tcp packet
+    plen = enc28j60PacketReceive(buffer_size, buf);
+    dat_p=packetloop_icmp_tcp(buf,plen);
+    if(dat_p==0) {
+      check_for_dhcp_answer( buf, plen);
+      dhcpState = dhcp_state();
+      // we are idle here
+      if( dhcpState != DHCP_STATE_OK ) {
+        if (HAL_GetTick() > (lastDhcpRequest + 10000L) ){
+          lastDhcpRequest = HAL_GetTick();
+	  if( --dhcpTries <= 0 )
+		  return 0;		// Failed to allocate address
+          // send dhcp
+          dhcp_start( buf, mymac, myip, mynetmask,gwip, dnsip, dhcpsvrip );
+        }
+      } else {
+        if( !gotIp ) {
+          gotIp = TRUE;
+
+          //init the ethernet/ip layer:
+          init_ip_arp_udp_tcp(mymac, myip, myport);
+
+          // Set the Router IP
+          client_set_gwip(gwip);  // e.g internal IP of dsl router
+
+#ifdef DNS_client
+          // Set the DNS server IP address if required, or use default
+          dnslkup_set_dnsip( dnsip );
+#endif
+
+        }
+      }
+    }
+  }
+
+  return 1;
+
+}
+
+// Perform all processing to resolve a hostname to IP address.
+// Returns 1 for successful Name resolution, 0 otherwise
+uint8_t resolveHostname(uint8_t *buf, uint16_t buffer_size, uint8_t *hostname ) {
+  uint16_t dat_p;
+  int plen = 0;
+  long lastDnsRequest = HAL_GetTick();
+  uint8_t dns_state = DNS_STATE_INIT;
+  bool gotAddress = FALSE;
+  uint8_t dnsTries = 3;	// After 10 attempts fail gracefully so other action can be carried out
+
+  while( !gotAddress ) {
+    // handle ping and wait for a tcp packet
+    plen = enc28j60PacketReceive(buffer_size, buf);
+    dat_p=packetloop_icmp_tcp(buf,plen);
+
+    // We have a packet
+    // Check if IP data
+    if (dat_p == 0) {
+      if (client_waiting_gw() ) {
+        // No ARP received for gateway
+        continue;
+      }
+      // It has IP data
+      if (dns_state==DNS_STATE_INIT) {
+        dns_state=DNS_STATE_REQUESTED;
+        lastDnsRequest = HAL_GetTick();
+        dnslkup_request(buf,hostname);
+        continue;
+      }
+      if (dns_state!=DNS_STATE_ANSWER){
+        // retry every minute if dns-lookup failed:
+        if (HAL_GetTick() > (lastDnsRequest + 60000L) ){
+	  if( --dnsTries <= 0 )
+	    return 0;		// Failed to allocate address
+
+          dns_state=DNS_STATE_INIT;
+          lastDnsRequest = HAL_GetTick();
+        }
+        // don't try to use client before
+        // we have a result of dns-lookup
+
+        continue;
+      }
+    }
+    else {
+      if (dns_state==DNS_STATE_REQUESTED && udp_client_check_for_dns_answer( buf, plen ) ){
+        dns_state=DNS_STATE_ANSWER;
+        //client_set_wwwip(dnslkup_getip());
+        client_tcp_set_serverip(dnslkup_getip());
+	gotAddress = TRUE;
+      }
+    }
+  }
+
+  return 1;
+}
+
 // return 0 to just continue in the packet loop and return the position 
 // of the tcp/udp data if there is tcp/udp data part
 uint16_t packetloop_icmp_tcp(uint8_t * buf, uint16_t plen) {
